network_spatial_coherence/requirements.txt
matplotlib==3.8.3
memory_profiler==0.61.0
networkx==3.2.1
nodevectors==0.1.23
numpy==1.26.4
pandas==2.2.1
pecanpy==2.0.9
Pillow==10.2.0
pixelator==1.2.0
plotly==5.19.0
pymde==0.1.18
python_igraph==0.11.4
scienceplots==2.1.1
scikit_learn==1.3.1
scipy==1.12.0
seaborn==0.13.2
torch==2.1.0
umap-learn
